split tags/filters into two modules
this removes the need for a nested function
