+++ argonauts/templatetags/argonauts/__init__.py
@@ -1,0 +1,5 @@
+from __future__ import absolute_import
+
+from .filters import json
+from .tags import json as json_tag
+from .library import register
