+++ argonauts/templatetags/argonauts/library.py
@@ -1,0 +1,3 @@
+from django import template
+
+register = template.Library()
